chore(dep): move AnimatorHash into Runtime folder
Prepare to separate property drawer into its own file.
